requirements.txt
langchain
langchain-community
langchain-google-community
langchain-nvidia-ai-endpoints
duckduckgo-search
google-search-results
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
streamlit